combine host.Local.start into host.Local.__init__

+++ tests/testutils.py
@@ -7,7 +7,7 @@ TC_PATH = "host/target/debug/tinychain"
 PORT = 8702
 
 
-def start_host(name, clusters):
+def start_host(name, clusters=[], overwrite=True):
     port = PORT
     if clusters:
         port = tc.uri(clusters[0]).port()
@@ -29,20 +29,19 @@ def start_host(name, clusters):
 
         cluster_config += str(cluster_uri.path())
 
-        tc.write_cluster(cluster, cluster_config, True)
+        tc.write_cluster(cluster, cluster_config, overwrite)
         config.append(cluster_config)
 
     data_dir = "/tmp/tc/tmp/" + name
-    if os.path.exists(data_dir):
+    if overwrite and os.path.exists(data_dir):
         shutil.rmtree(data_dir)
 
-    host = tc.host.Local(
+    print(f"start host on port {port}")
+    return tc.host.Local(
+        TC_PATH,
         workspace="/tmp/tc/tmp/" + name,
         data_dir=data_dir,
         clusters=config,
+        port=port,
         force_create=True)
-
-    print(f"start host on port {port}")
-    host.start(TC_PATH, port, log_level="debug")
-    return host
